Put all Impellers into 1 Script + Rotation
I made a script to control all impellers directions so we just need one
asset for the impeller and then set its direction and it'll rotate the
sprite in-game. It'll always be shown as "Up" in the editor but it
rotates when the scene runs.

Also added rotation to the rover image so you can see the direction it's
going at.

//not sure about these meta files though....